Fix support for no verify
On RHEL 7, which has Stunnel version 4.56, we found that a different
configuration was required.

Instead of specifying verify = default, it is in fact necessary to
simply not specify verify at all.

Without this patch applied, an error message is seen:

  "verify = default": Bad verify level

On RHEL 6, Stunnel version 4.29 is available and it behaves differently,
but it also allows verify to simply be not specified.

Therefore, this patch modifies templates/stunnel.conf.erb to simply not
add the verify line if verify level default is specified.